설정/북마크 게시글 부분 파일 분리
