+++ src/app/profile/bookmark/BookmarkPage.scss
@@ -14,235 +14,7 @@
 
         width: 61.5%;
     }
-    .studyContainer{
-        display: flex;
-        flex-direction: row;
-        justify-content: space-between;
-        align-items: center;
-
-        width: 100%;
-    }
-    .title{
-        font-size: 24px;
-        line-height: 35px;
-
-        color: #1a1a1a;
-    }
-    .borderLine{
-        display: flex;
-        flex: 1;
-
-        border-bottom-width: 1px;
-        border-bottom-color: #c1c1c1;
-
-        margin-left: 34px;
-        
-    }
-    .showAllText{
-        color:#818181;
-        font-size: 16px;
-        line-height: 24px;
-
-        margin-left: 17px;
-    }
-    .bottomTriangleButton{
-        // width: 0px;
-        // height: 0px;
-
-        // border-left-width: 7.5px;
-        // border-right-width: 7.5px;
-        // border-top-width: 15px;
-        // border-style: solid;
-
-        // background-color: transparent;
-        // border-left-color: transparent;
-        // border-right-color: transparent;
-        // border-top-color: red;
-
-        width: 0; 
-        height: 0; 
-        border-left: 10px solid transparent;
-        border-right: 10px solid transparent;
-        
-        border-top: 15px solid #f00;
-    }
-
-
-// studyList
-    .studyList{
-        display: flex;
-        flex-direction: row;
-        flex:1;
-
-        justify-content: space-between;
-
-        margin-top: 28px;
-        // background-color: red;
-    }
-    .studyItemContainer{
-        // width: 18.69%;
-        // height: 15.9%;
-        display: flex;
-        flex-direction: column;
-        background-color: blue;
-        border-radius: 10px;
-
-        flex:1;
-
-        // width: 220px;
-        // height: 220px;
-    }
-    .studyItemName{
-        font-size: 12px;
-        line-height: 17px;
-        color: #ffffff;
-        margin-top: 14px;
-        margin-left: 16px;
-    }
-    .studyItemTitle{
-        font-size: 24px;
-        color:#ffffff;
-        margin-top: 24px;
-        margin-left: 16px;
-        margin-right: 16px;
-        line-height: 35px;
-    }
-    .studyItemInfoContainer{
-        margin-left: 12px;
-        margin-right: 12px;
-        margin-top: 28px;
-
-        display: flex;
-        flex-direction: row;
-        justify-content: space-between;
-    }
-    .studyItemInfo{
-        font-size: 10px;
-        line-height: 15px;
-        color:#ffffff;
-
-    }
-    .tagContainer{
-        background-color: #ffffff;
-
-        display: flex;
-        flex-direction: row;
-        height: 48px;
-        align-items: center;
-
-        margin-top: 4px;
-        padding-left: 7px;
-
-        // box-shadow: none | x-position y-position blur spread color | inset | initial | inherit
-        box-shadow: 1px 2px 4px 0px rgba(95, 95, 95, 0.16);
-    }
-    .tagItemText{
-        color:#6f80ff;
-        font-size: 12px;
-        line-height: 10px;
-
-        padding: 4px 17px;
-        margin-right: 4px;
-
-        border: solid;
-        border-radius: 12px;
-        border-width: 1px;
-        border-color: #6f80ff;
-    }
-
-
-    //articleList
-    .articleHeaderContainer{
-        width: 100%;
-        height: 45px;
-
-        display: flex;
-        flex-direction: row;
-        flex:1;
-
-        justify-content: space-between;
-        align-items: center;
-
-        background-color: #f7f8ff;
-
-        margin-top: 20px;
-        
-    }
-    .articleInfoContainer{
-        display: flex;
-        flex-direction: row;
-
-        flex:3;
-    }
-    .articleText{
-        font-size: 16px;
-        line-height: 24px;
-        color: #1a1a1a;
-
-    }
-    .articleTitle{
-        flex:7;
-        // flex-shrink: 2;
-        overflow:hidden; //넘치는 경우 숨김
-        text-overflow: ellipsis;
-        white-space:nowrap; //한줄로 나오게 함
-
-    }
-    .articleInfo{
-        display: flex;
-        flex:1;
-        // background-color: red;
-
-        // border-right: solid;
-        // border-right-width: 2px;
-        // border-right-color: blue;
-
-        // align-self: center;
-        justify-content: center;
-        align-items: center;
-    }
-
-    .articleContainer{
-        width: 100%;
-        height: 64px;
-
-        display: flex;
-        flex-direction: row;
-
-        justify-content: space-between;
-        align-items: center;
-
-        background-color: #ffffff;
-
-        border-bottom: solid;
-        border-bottom-color: #e0e0e0;
-        border-bottom-width: 1px;
-
-        // margin-top: 20px;
-    }
-
-
-    .paginationContainer{
-        display: flex;
-        flex-direction: row;
-
-        justify-content: space-between;
-        align-self: center;
-
-        width: 20%;
-
-        margin-bottom: 50px;
-    }
-    .paginationText{
-        @include paginationButton(#05158b)
-    }
-    .paginationNumber{
-        @include paginationButton(#707070)
-    }
-
-    button:active{
-        opacity: 0.3;
-    }
+    
 }
 
 @include tablet{}
