+++ src/app/profile/bookmark/article/ArticleHeader.tsx
@@ -1,0 +1,13 @@
+import React from 'react'
+import './BookmarkArticle.scss'
+
+const ArticleHeader=()=>{
+    return(
+        <div className='studyContainer'>
+            <div className='title'>게시글</div>
+            <hr className='borderLine'/>
+        </div>
+    )
+}
+
+export default ArticleHeader
