+++ src/app/profile/bookmark/study/BookmarkStudy.scss
@@ -8,7 +8,38 @@
 
 @include desktop{
 
+    .studyContainer{
+        display: flex;
+        flex-direction: row;
+        justify-content: space-between;
+        align-items: center;
 
+        width: 100%;
+    }
+    .title{
+        font-size: 24px;
+        line-height: 35px;
+
+        color: #1a1a1a;
+    }
+    .borderLine{
+        display: flex;
+        flex: 1;
+
+        border-bottom-width: 1px;
+        border-bottom-color: #c1c1c1;
+
+        margin-left: 34px;
+        
+    }
+    .showAllText{
+        color:#818181;
+        font-size: 16px;
+        line-height: 24px;
+
+        margin-left: 17px;
+    }
+    
 // studyList
     .studyList{
         display: flex;
